+++ shared/common/src/main/java/cz/inovatika/dochub/impl/FilePermanentContentSpaceImpl.java
@@ -42,8 +42,8 @@ public class FilePermanentContentSpaceImpl implements PermanentContentSpace, Cle
     }
 
     @Override
-    public int getConfiguredMaxLimit() {
-        int maxGb = KConfiguration.getInstance().getConfiguration().getInt("dochub.permanent.max.size.gb", 10);
+    public double getConfiguredMaxLimit() {
+        double maxGb = KConfiguration.getInstance().getConfiguration().getDouble("dochub.permanent.max.size.gb", 10);
         return maxGb;
     }
